+++ R/peaks2spec.R
@@ -46,7 +46,7 @@ peaks2spec <- function(recording, sound_peaks, spec_size = NA,
   tx <- recording$tx
   input_shape <- c(
     spec_size / time_step_size,
-    (freq_range[2] - freq_range[1]) * window_length
+    ((freq_range[2] - freq_range[1]) * frequency_resolution) * window_length
   )
 
   windows <- matrix(NA,
@@ -92,7 +92,6 @@ peaks2spec <- function(recording, sound_peaks, spec_size = NA,
     )
     spec2_filt <- spec2[, -filt]
     spec2_filt <- spec2_filt + dynamic_range
-    dim(spec2_filt)
 
     fmaxe <- dimnames(spec2_filt)[[2]][which(spec2_filt == max(spec2_filt), arr.ind = TRUE)[2]]